Force uninstalling and removing printers and downloaded drivers
Printers configured in the system will not notice the 32bit > 64bit
as long as they can still find the PPD in /etc/cups/ppd that was
used to configure them in 2.6, which will cause trouble if the
underlying driver has not been automatically migrated to 64bit.

This will be a problem particularly for drivers downloaded from
OpenPrinting.org via eos-config-printer, as there's no easy way
to automatically update them in the system, but could also be
a problem for other printers supported via OS upgrades if they
have 32bit drivers available but not 64bit ones (e.g. all Canon
printers supported by the cnijfilter-common driver), which will
obscurely stop working after the upgrade, even if they will show
up as installed to the user (as they will keep the PPD file).

To prevent the trouble and the confusion, and even though is a
bit of a drastic measure, we will uninstall every printer and
every downloaded driver from the system, so that the user is
forced to manually install the printers again upon reinstallation.

https://phabricator.endlessm.com/T13900