Don't hide the error output of printers related commands
We had a report from an user that got this script failing when stopping
the CUPS service because of the -e flag, and we don't know why that was
since the error output was being redirected to /dev/null.

As the script is run with -e, let's remove this silly idiom and let
the script fail and report the problem in all its glory.

https://phabricator.endlessm.com/T19677