Don't stop CUPS to remove printers and printing jobs
It seems this worked in the past because of a bug that would cause
the cups.socket and cups.path services enabled even after stopping
CUPS, meaning that in the end the CUPS daemon would run anyway when
executing `lpstat`, `cancel` and `lpadmin`, which are commands that
won't work fine otherwise.

So, it seems the whole Stop -> remove printers/jobs -> Start CUPS
dance is not correct after all, and all we probably need, if anything,
is just to restart CUPS before removing things to make sure the list
is up-to-date, and also after to make sure things are correctly updated
after using `cancel` and `lpadmin`.

https://phabricator.endlessm.com/T19677